+++ lib/rubocop/cop/team.rb
@@ -12,7 +12,8 @@ module RuboCop
         Style::LineEndConcatenation => [Style::UnneededInterpolation],
         Style::UnneededInterpolation => [Style::LineEndConcatenation],
         Style::SelfAssignment => [Style::SpaceAroundOperators],
-        Style::SpaceAroundOperators => [Style::SelfAssignment]
+        Style::SpaceAroundOperators => [Style::SelfAssignment],
+        Style::BracesAroundHashParameters => [Style::MultilineHashBraceLayout]
       }.freeze
 
       DEFAULT_OPTIONS = {